Use volley for internet calls
