+++ app/src/main/java/br/bosseur/popuplarmoviesapp/listeners/TaskListener.java
@@ -1,0 +1,26 @@
+package br.bosseur.popuplarmoviesapp.listeners;
+
+/**
+ * Interface for doing callbacks in stages calls
+ * on internet connection
+ */
+
+public interface TaskListener<T> {
+    /**
+     * Should be called when an internet call starts
+     */
+    void onStartTask();
+
+    /**
+     * Should be called when an error occurs.
+     * @param errorMessage Explanation of the error that occurred.
+     */
+    void onError(String errorMessage);
+
+    /**
+     * Should be called when the call to the internet returns
+     * passing the data that was received.
+     * @param data The data that was returned by the call
+     */
+    void onCompleteTask(T data);
+}
