--- app/src/main/java/br/bosseur/popuplarmoviesapp/tasks/AsyncTaskListener.java
@@ -1,11 +1,0 @@
-package br.bosseur.popuplarmoviesapp.tasks;
-
-/**
- * Created by EKoetsier on 29/01/2018.
- */
-
-public interface AsyncTaskListener<T> {
-    void onStartTask();
-    void onError(String errorMessage);
-    void onCompleteTask(T data);
-}
