+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -33,6 +33,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="?attr/actionBarSize"
                 app:layout_collapseMode="pin"
+                android:theme="@style/ToolbarStyle"
                 />
 
         </android.support.design.widget.CollapsingToolbarLayout>
