+++ app/src/main/res/layout/activity_movies.xml
@@ -10,6 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="?attr/colorPrimary"
+        android:theme="@style/ToolbarStyle"
         app:titleTextColor="@android:color/white">
 
     </android.support.v7.widget.Toolbar>
